# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Shoko Image Planner Plugin API
  version: '1.0.0'
  description: Versioned, admin-only image planning API for Nakama and other clients.
servers:
  - url: /
security:
  - shokoApiKey: []
paths:
  /api/v3/Plugin/ImagePlanner/status:
    get:
      tags: [ImagePlanner]
      responses:
        '200': {description: Status, content: {application/json: {schema: {$ref: '#/components/schemas/Status'}}}}
  /api/v3/Plugin/ImagePlanner/capabilities:
    get:
      tags: [ImagePlanner]
      responses:
        '200': {description: Capabilities, content: {application/json: {schema: {$ref: '#/components/schemas/Status'}}}}
  /api/v3/Plugin/ImagePlanner/providers:
    get:
      tags: [ImagePlanner]
      responses:
        '200': {description: Providers, content: {application/json: {schema: {type: array, items: {$ref: '#/components/schemas/Provider'}}}}}
  /api/v3/Plugin/ImagePlanner/groups:
    get:
      tags: [ImagePlanner]
      responses:
        '200': {description: Top-level groups, content: {application/json: {schema: {type: array, items: {$ref: '#/components/schemas/Group'}}}}}
  /api/v3/Plugin/ImagePlanner/plan:
    post:
      tags: [ImagePlanner]
      parameters: [{$ref: '#/components/parameters/PlanIdempotencyKey'}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/PlanRequest'}}}}
      responses:
        '200': {description: Plan, content: {application/json: {schema: {$ref: '#/components/schemas/PlanResponse'}}}}
        '400': {$ref: '#/components/responses/Problem'}
  /api/v3/Plugin/ImagePlanner/apply:
    post:
      tags: [ImagePlanner]
      parameters: [{$ref: '#/components/parameters/IdempotencyKey'}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/MutationRequest'}}}}
      responses:
        '200': {description: Applied plan, content: {application/json: {schema: {$ref: '#/components/schemas/PlanResponse'}}}}
        '400': {$ref: '#/components/responses/Problem'}
  /api/v3/Plugin/ImagePlanner/reconcile:
    post:
      tags: [ImagePlanner]
      parameters: [{$ref: '#/components/parameters/IdempotencyKey'}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/MutationRequest'}}}}
      responses:
        '200': {description: Reconciled plan, content: {application/json: {schema: {$ref: '#/components/schemas/PlanResponse'}}}}
        '400': {$ref: '#/components/responses/Problem'}
components:
  securitySchemes:
    shokoApiKey: {type: apiKey, in: header, name: apikey}
  parameters:
    PlanIdempotencyKey: {name: Idempotency-Key, in: header, required: false, description: Required when the request sets ingest to true., schema: {type: string, maxLength: 200}}
    IdempotencyKey: {name: Idempotency-Key, in: header, required: true, schema: {type: string, maxLength: 200}}
  responses:
    Problem: {description: RFC 7807 problem, content: {application/problem+json: {schema: {$ref: '#/components/schemas/Problem'}}}}
  schemas:
    Status:
      type: object
      required: [apiVersion, pluginVersion, minimumShokoAbstractionsVersion, enabled, capabilities]
      properties:
        apiVersion: {type: string, example: '1'}
        pluginVersion: {type: string}
        minimumShokoAbstractionsVersion: {type: string}
        enabled: {type: boolean}
        capabilities: {type: array, items: {$ref: '#/components/schemas/Capability'}}
    ApiVersion: {type: object, required: [apiVersion, pluginVersion, minimumShokoAbstractionsVersion], properties: {apiVersion: {type: string}, pluginVersion: {type: string}, minimumShokoAbstractionsVersion: {type: string}}}
    Capability: {type: object, required: [name, enabled], properties: {name: {type: string}, enabled: {type: boolean}, detail: {type: string, nullable: true}}}
    Provider: {type: object, required: [name, source, configured, priority], properties: {name: {type: string}, source: {type: string}, configured: {type: boolean}, priority: {type: integer}}}
    Group: {type: object, required: [id, name, seriesCount], properties: {id: {type: integer}, name: {type: string}, seriesCount: {type: integer}}}
    PlanRequest: {type: object, properties: {apiVersion: {type: integer, default: 1}, groupIds: {type: array, items: {type: integer}, nullable: true}, ingest: {type: boolean, default: false, description: When true, Idempotency-Key is required and selected provider images are downloaded and stored.}, force: {type: boolean, default: false}}}
    MutationRequest: {type: object, properties: {apiVersion: {type: integer, default: 1}, groupIds: {type: array, items: {type: integer}, nullable: true}, force: {type: boolean, default: false}}}
    Assignment: {type: object, required: [seriesId, candidateId, imageType, isUnique, isFallback, score], properties: {seriesId: {type: integer}, candidateId: {type: string}, imageType: {type: string}, isUnique: {type: boolean}, isFallback: {type: boolean}, score: {type: integer, format: int64}, reason: {type: string, nullable: true}}}
    SeriesPlan: {type: object, required: [seriesId, name, assignments, protected], properties: {seriesId: {type: integer}, name: {type: string}, assignments: {type: array, items: {$ref: '#/components/schemas/Assignment'}}, protected: {type: boolean}, protectedReason: {type: string, nullable: true}}}
    GroupPlan: {type: object, required: [groupId, name, series, hasInsufficientUniqueCandidates], properties: {groupId: {type: integer}, name: {type: string}, series: {type: array, items: {$ref: '#/components/schemas/SeriesPlan'}}, hasInsufficientUniqueCandidates: {type: boolean}}}
    PlanResponse: {type: object, required: [apiVersion, createdAt, groups, downloadsPerformed, changedCount], properties: {apiVersion: {type: integer}, createdAt: {type: string, format: date-time}, groups: {type: array, items: {$ref: '#/components/schemas/GroupPlan'}}, downloadsPerformed: {type: boolean}, changedCount: {type: integer}}}
    Problem: {type: object, properties: {type: {type: string}, title: {type: string}, status: {type: integer}, detail: {type: string}, instance: {type: string}}}